feat(repo): manual run script
