+++ .github/workflows/build.yml
@@ -26,7 +26,6 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: volta-cli/action@v1
-      - run: yarn install --frozen-lockfile --production
       - run: yarn run build:ci
       - run: yarn run build-az-task
         env:
