+++ README.md
@@ -25,7 +25,7 @@ Contents:
 - `checkcommit` - validate commit message
 - `clean` - remove coverage data, Jest cache and transpiled files,
 - `clean-deps` - remove `node_modules`
-- `clean-lib` - remove coverage and build directories and files
+- `clean-generated` - remove coverage and build directories and files
 - `clean-logs` - remove log files
 - `commit` - start commit
 - `format` - format files
@@ -34,6 +34,55 @@ Contents:
 - `sort-pj` - sort `package.json`
 - `test` - run tests,
 - `test:watch` - interactive watch mode to automatically re-run tests
+- `manual-run` - invoke the work item association api locally
+
+## Building
+
+To build the project with yarn:
+
+```bash
+yarn install
+yarn build
+```
+
+### Building the Azure Devops Task
+
+To build the Azure Devops task with yarn:
+
+```bash
+yarn build:ci
+yarn build-az-task
+```
+
+The Task build will output a `.vsix` file that can be uploaded to Azure Devops
+
+## Running
+
+### Manual run
+
+To invoke the task locally, you will need to:
+
+1. Copy the [sample.env](sample.env) and rename it `.env`
+2. Populate the values _blank_ values in your new `.env` file.
+   - `AZURE_PERSONAL_ACCESS_TOKEN` is a valid Personal Access Token([PAT])
+   - `ORG_URL` is the URL for your organization: e.g. `https://dev.azure.com/cool-co`
+   - `PROJECT` is your project name. e.g. `https://dev.azure.com/cool-co/cool-code`
+3. Lastly, you need a valid build id to execute against. The build id is an integer value that
+   can be found with the [predefined variable] `$(Build.BuildId)`.
+
+To execute the task with yarn:
+
+```bash
+yarn run manual-run SOME_BUILD_ID
+```
+
+## Testing
+
+To run tests with yarn:
+
+```bash
+yarn test
+```
 
 [typescript]: https://www.typescriptlang.org/
 [typescript-4-0]: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-0.html
@@ -44,3 +93,5 @@ Contents:
 [volta-getting-started]: https://docs.volta.sh/guide/getting-started
 [dependabot]: https://dependabot.com/
 [kodiakhq]: https://kodiakhq.com/
+[pat]: https://docs.microsoft.com/en-us/azure/devops/organizations/accounts/use-personal-access-tokens-to-authenticate?view=azure-devops&tabs=preview-page#create-a-pat
+[predefined variable]: https://docs.microsoft.com/en-us/azure/devops/pipelines/build/variables?view=azure-devops&tabs=yaml
