+++ package.json
@@ -10,7 +10,7 @@
     "build-az-task": "npx tfx extension create --manifest-globs vss-extension.json",
     "prebuild:ci": "run-s clean",
     "build:ci": "tsc -p tsconfig.release.json",
-    "postbuild:ci": "cp package.json task/ && cp -r node_modules/ task/node_modules && cp task.json task/",
+    "postbuild:ci": "cp package.json task/ && cp yarn.lock task/ && cp task.json task/ && cd task && yarn install --production --frozen-lockfile",
     "build:watch": "tsc -w -p tsconfig.release.json",
     "checkcommit": "node ./scripts/commit-lint.js",
     "clean": "run-p clean-generated clean-logs",
@@ -20,6 +20,7 @@
     "commit": "git-cz",
     "format": "./scripts/format.sh && yarn run sort-pj",
     "lint": "scripts/check-format.sh",
+    "manual-run": "NODE_ENV=local run-func task/index.js main",
     "nuke": "run-s clean clean-deps",
     "sort-pj": "npx sort-package-json",
     "test": "jest",
@@ -81,6 +82,7 @@
     "prettier": "~2.1.2",
     "pretty-quick": "^3.0.2",
     "rimraf": "~3.0.2",
+    "run-func": "^1.0.5",
     "satisfied": "^1.1.2",
     "sort-package-json": "^1.44.0",
     "sync-request": "^6.1.0",
