+++ sample.env
@@ -1,0 +1,3 @@
+AZURE_PERSONAL_ACCESS_TOKEN=BLANK
+ORG_URL=BLANK
+PROJECT=BLANK
